class Dependencies
added functions
